return main file and config main.lambda
