+++ src/main.ts
@@ -1,12 +1,12 @@
 import { NestFactory } from '@nestjs/core';
-import serverlessExpress from '@codegenie/serverless-express';
-import { Callback, Context, Handler } from 'aws-lambda';
-import { AppModule } from './app.module';
+
 import helmet from 'helmet';
 
-let server: Handler;
+import { AppModule } from './app.module';
 
-async function bootstrap(): Promise<Handler> {
+const port = process.env.PORT || 4000;
+
+async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
   app.enableCors({
@@ -14,17 +14,8 @@ async function bootstrap(): Promise<Handler> {
   });
   app.use(helmet());
 
-  await app.init();
-
-  const expressApp = app.getHttpAdapter().getInstance();
-  return serverlessExpress({ app: expressApp });
+  await app.listen(port);
 }
-
-export const handler: Handler = async (
-  event: any,
-  context: Context,
-  callback: Callback,
-) => {
-  server = server ?? (await bootstrap());
-  return server(event, context, callback);
-};
+bootstrap().then(() => {
+  console.log('App is running on %s port', port);
+});
